requirements.txt
asgiref==3.5.2
Django==4.0.6
django-ckeditor==6.4.2
django-debug-toolbar==3.5.0
django-js-asset==2.0.0
django-modeltranslation==0.18.4
djangorestframework==3.13.1
Pillow==9.2.0
psycopg2==2.9.3
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
Unidecode==1.3.4
